main.py: Convert image URLs to absolute URLs
As recommended in many places, and to have canonical URLs, this commit
parses the generated templates before serving them to the client. In this
process, the server will now look for any image in the form:

<img prop1="1" src="image1.png" prop2="2"></img>

And it will subsitute the URL with the canonical URL.

In order to have the code structured and that the functions used keep the
same meaning, `template.render_template()` has been added, which is
basically a wrapper around `flask.render_template()`.